Actualizacion del Readme y el back

+++ controllers/productoController.js
@@ -63,28 +63,19 @@ export const createProducto = async (req, res) => {
     try {
         const product = await Producto.findOne({ nombre: req.body.nombre, categoria: req.body.categoria });
         if (product) {
-            /*if (req.files?.image) {
+            if (req.files?.image) {
                 await fs.unlink(req.files.image.tempFilePath);
-            }*/
+            }
             return res.render("errorAdmin", { error: "producto ya registrado"})
         }
         const productoNuevo =await Producto.create(req.body);
         console.log(req.files.image)
         if (req.files?.image) {
-            /* const tempFilePath = req.files.image.tempFilePath
-            const originalName = req.files.image.name
-            const fileExtencion = originalName.split('.').pop()*/
             const productImg = await analizarCategoria(productoNuevo,req);
             productoNuevo.imagen = {
                 public_id: productImg.public_id,
                 secure_url: productImg.secure_url
             }
-            /* const newTempFilePath = `${req.files.image.tempFilePath}.${fileExtencion}`
-            await fs.rename(tempFilePath, newTempFilePath, (err) => {
-                if (err) {
-                    console.log(err)
-                }
-            })*/
             await fs.unlink(req.files.image.tempFilePath);
         } else {
             res.render('errorAdmin', {
@@ -114,22 +105,12 @@ export const createProductos = async (req, res) => {
         }
         const productoNuevo = Producto.insertMany(req.body);
         if (req.files?.image) {
-            // const tempFilePath = req.files.image.tempFilePath
-            // const originalName = req.files.image.name
-            // const fileExtencion = originalName.split('.').pop()
-            let productImg=analizarCategoria(req.body,req)
+
+            let productImg=await analizarCategoria(req.body,req)
         productoNuevo.imagen = {
             public_id: productImg.public_id,
             secure_url: productImg.secure_url
         }
-        // const newTempFilePath = `${req.files.image.tempFilePath}.${fileExtencion}`
-        // await fs.rename(tempFilePath, newTempFilePath, (err) => {
-        //     if (err) {
-        //         res.render('error', {
-        //             error: 'no se subio la imagen'
-        //         })
-        //     }
-        // })
         await fs.unlink(req.files.image.tempFilePath);
         await productoNuevo.save();
         const producto=await Producto.findById(productoNuevo._id).lean();
@@ -143,22 +124,6 @@ export const createProductos = async (req, res) => {
         })
     }
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 export const deleteProducto = async (req, res) => {
     try {
         const productoId = req.params.id;
@@ -197,50 +162,25 @@ export const viewUpdateProduct=async (req,res)=>{
 export const updateProducto = async (req, res) => {
 
     try {
-
-
         const productoId = req.params.id;
-
-
         const productoActualizado = await Producto.findById(productoId, req.body);
         if (!productoActualizado) {
-
-
             return res.status(404).json({ error: "producto no encontrado" });
         }
-
-
         const productoNuevo = await Producto.findByIdAndUpdate(productoId, req.body, { new: true });
         if (req.files?.image) {
-
-
             if(productoNuevo.imagen.public_id){
                 console.log(productoNuevo.imagen.public_id)
-
-
                 await deletePhoto(productoNuevo.imagen.public_id);
-
             }
-            // const tempFilePath = req.files.image.tempFilePath
-            // const originalName = req.files.image.name
-            // const fileExtencion = originalName.split('.').pop()
             let productImg=await analizarCategoria(productoNuevo,req)
             productoNuevo.imagen = {
                 public_id: productImg.public_id,
                 secure_url: productImg.secure_url
             }
-            // const newTempFilePath = `${req.files.image.tempFilePath}.${fileExtencion}`
-            // await fs.rename(tempFilePath, newTempFilePath, (err) => {
-            //     if (err) {
-            //         res.render('error',{
-            //             error:err
-            //         })
-            //     }
-            // })
             await fs.unlink(req.files.image.tempFilePath);
             await productoNuevo.save();
         } 
-
         const producto=await Producto.findById(productoId).lean();
         res.render('productoIdAdmin', {
             producto
